Fix pio2 issue with vars sharing iodescs
If multiple variables share the same iodesc, pio2 currently
assumes that all variables have the same layout including the
same number of dimensions. This is an issue when two variables
differ only the unlimited/time dimension but share the same
iodesc.

In this commit we make sure that the startlist and countlist
passed to pnetcdf is correct by modifying it for each variable
(based on whether the variable has an unlimited dim or not).