bug fix for ppool blow up during spinup
Now supplemental P is added to vegetation only when ppool
is insufficient to support plant P demand during AD spinup.
This avoids a numerical blowup in 2-stage AD spinup.
BeTR is also updated accordingly for this bug fix.
A bug is also fixed in pftvarcon in check read in parameter
values of harvestprod10.